Feat : DTO Fix
Former-commit-id: 16f8a5e264c2a0445a5b7e929cfc9e1b407510ff

+++ backend/src/main/java/com/a503/churros/repository/news/LikeRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 public interface LikeRepository extends JpaRepository<Like , Long> {
     Optional<List<Like>> findByUserIdx(long userIdx);
-    Optional<Like> findByUserIdxAndArticleIdx(long userIdx , int articleIdx);
+    Optional<Like> findByUserIdxAndArticleIdx(long userIdx , long articleIdx);
 }
