+++ backend/src/main/java/com/a503/churros/service/news/NewsService.java
@@ -8,10 +8,10 @@ import java.util.List;
 public interface NewsService {
     List<Integer> sendRecommend(long userId);
     List<Integer> sendSample();
-    void saveReadArti(long userId, int articleId);
-    void recordLike(long userId, int articleId);
+    void saveReadArti(long userId, long articleId);
+    void recordLike(long userId, long articleId);
     List<Long> getLikeList(long userIdx);
-    void recordDisLike(long userId, int articleId);
+    void recordDisLike(long userId, long articleId);
 
     ArticleDTO getArticleInfo(/*long userId , */long articleId);
     String callNaver(String url);
